Implement selectable IP resolver in options
Extention largely overhauled as a result of this new feature. Options
page allows user to select and order their preferred IP sources. Also
includes field for 1 manual IP source entry. If an IP is not found from
the first source, fall-back to the second, and so on. Text area in
bubble has small status updates during this fall-back process.