Drop unused messaging features
Storage change handling isn't worth the effort. We already treat the
storage area as the source of truth, so the only benefit would be to
trigger a notification on the options page that "settings have changed!"
if a user happens to have multiple computers with the options page open.
Let it go.